Add core.tim with draft of inputomata function
This is a rough, proof-of-concept finite state machine generator that
accepts user key presses, and navigates the FSM based on them. Mappings
from keys in each state to other states (by name) are defined in a very
simple map format. This keeps the plugin writer (or hobbyist) from
needing to define any Vim key mappings, while still allowing the total
takeover of the keyboard for any length of time. Right now it can hop
states and allow the user to exit, and it also can print an optional
message per state. Soon I want it to be able to call functions and
actually affect the world.