Remove core.tim/inputomata states defaults merging
There appears to be a bug with TimL's conj. I've submitted an issue on
github, but in the meantime, I've pulled out the merging of the
defaults, and state-hopping and end conditions work nicely again. The
idea with defaults merging was to allow the user to skip filling in the
:start and :end states. I'd provide a default up front, e.g.:

    {:start {:msg "Press q to quit"
             \q :end}
     :end nil}

Then passed-in states map would conj onto that, overwriting those
entries, if provided. This would allow leaving out either, or both
states. It also serves as a demo function; call it with an empty map,
and you're put in an FSM that traps key presses (and the user) until the
q key is pressed, and it informs the user of this by default.