Fill out inputomata use case in core.tim
The need for which I've wanted to have this finite state mechanism in
TimL is a little in-place color manipulator. I could move through rgb
components with, say, h and l, and push each value up and down with j
and k. Here's a more fleshed-out example of something like that.