Fix inputomata in core.tim, with flowchart help
I drew out a flowchart on my whiteboard, and it helped me finally get
all the logic in place. I also added a check for the [optional] :msg in
each state to see if it's a string; if it is, it prints it, and if not,
it prints the evaluation of it, which allows you to write messages with
values in them from elsewhere. The msg bit also clears out the status
line before this, and adds "\r" to each print to rewind to col 0.