Create plugin/inputomata.vim; organize tim files
This is tested and working. I can launch Vim, navigate to the color.tim
example file, and `:source %` or `ggcpG` to launch the color FSM. For
some reason, sourcing it inhibits the "done." message when quitting out
of the state machine through the provided \q state transition; the final
text displayed before pressing q is left in the status line.