Extract inputomata message printing
When run manually from within a TimL file (e.g. via cpip), the final
string would print out, but when run as a plugin, the terminal message
would never print. Some Vim help research suggested that lazy redrawing
may be to blame. The answer is to call redraw first to force it. This is
now extracted out to a status function.